sua lai kieu du lieu numberOfBookByPrice

+++ src/Operation.java
@@ -30,7 +30,7 @@ public class Operation {
         System.out.println("Số lượng sách Fiction Books thuộc thể loại 'Viễn tưởng 1' là: " + numberOfBookByCategory);
 
         //  Đếm số sách Fiction Books có giá nhỏ hơn 100
-        double numberOfBookByPrice = countFictionBooksByPrice(listBooks, 100);
+        int numberOfBookByPrice = countFictionBooksByPrice(listBooks, 100);
         System.out.println("Số lượng sách Fiction Books có giá nhỏ hơn 100 là: " + numberOfBookByPrice);
         //  Tìm kiếm giá của cuốn sách có tên được nhập vào từ bàn phím
